Bonsai_2.0/requirements.txt
Flask==2.3.2
websocket-client==1.5.2
pandas==2.1.1
numpy==1.26.0
plotly==5.17.0
scipy==1.11.3
